requirements.txt
# API Gateway requirements

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0

# GraphQL
strawberry-graphql[fastapi]==0.219.0

# HTTP Client
httpx==0.26.0

# Environment
python-dotenv==1.0.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2
